structure refactoring for wpf project
