+++ EasySubtitle.Console/Program.cs
@@ -1,11 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading;
-using System.Threading.Tasks;
 using EasySubtitle.WPF;
+using EasySubtitle.WPF.Windows;
 
 namespace EasySubtitle.Console
 {
