+++ EasySubtitle.WPF/ViewModels/ProgressDialogViewModel.cs
@@ -1,13 +1,10 @@
 using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.IO;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Windows.Input;
 
-namespace EasySubtitle.WPF
+namespace EasySubtitle.WPF.ViewModels
 {
     public class ProgressDialogViewModel : ViewModelBase
     {
@@ -192,128 +189,5 @@ namespace EasySubtitle.WPF
         }
 
         #endregion
-    }
-
-    public class SearchAdvancedSubtitleViewModel : ViewModelBase
-    {
-        private SelectedFile _selectedFile;
-
-        public SearchAdvancedSubtitleViewModel(IEnumerable<string> selectedFilePaths)
-        {
-            if (selectedFilePaths == null || !selectedFilePaths.Any()) throw new ArgumentNullException("selectedFilePaths");
-
-            SelectedFiles = new List<SelectedFile>();
-
-            selectedFilePaths.ToList().ForEach(x => SelectedFiles.Add(new SelectedFile(x)));
-            SelectedFile = SelectedFiles.FirstOrDefault();
-
-            Download = new DelegateCommand(() =>
-            {
-                DownlaodSubtitles();
-                var model = new ProgressDialogViewModel();
-                var progress = new ProgressDialogWindow {DataContext = model};
-                progress.Show();
-            });
-        }
-
-        private void DownlaodSubtitles()
-        {
-            SelectedFiles.ToList().Where(x => !x.Subtitles.Any()).ToList().ForEach(x => x.CheckSubtitles());
-
-            //download.. call download service etc.
-        }
-
-        public IList<SelectedFile> SelectedFiles { get; set; }
-
-        public SelectedFile SelectedFile
-        {
-            get { return _selectedFile; }
-            set
-            {
-                base.RaisePropertyChangingEvent("SelectedFile");
-                _selectedFile = value;
-                _selectedFile.CheckSubtitles();
-                base.RaisePropertyChangedEvent("SelectedFile");
-            }
-        }
-
-        public ICommand Download { get; set; }
-    }
-
-    public class DelegateCommand : ICommand
-    {
-        private readonly Action _action;
-
-        public DelegateCommand(Action action)
-        {
-            _action = action;
-        }
-
-        public bool CanExecute(object parameter)
-        {
-            return true;
-        }
-
-        public void Execute(object parameter)
-        {
-            if (_action != null)
-                _action.Invoke();
-        }
-
-        public event EventHandler CanExecuteChanged;
-    }
-
-    public class SelectedFile
-    {
-        public SelectedFile()
-        {
-
-        }
-
-        public SelectedFile(string filePath)
-        {
-            if (string.IsNullOrWhiteSpace(filePath))
-                throw new ArgumentNullException("filePath");
-
-            File = filePath;
-            Checked = true;
-            Subtitles = new List<FoundSubtitle>();
-
-            try
-            {
-                FileName = Path.GetFileName(filePath);
-            }
-            catch (Exception)
-            {
-                throw new InvalidOperationException("File path is not correct.");
-            }
-        }
-
-        public string File { get; set; }
-        public string FileName { get; set; }
-        public bool Checked { get; set; }
-        public IList<FoundSubtitle> Subtitles { get; set; }
-
-        public void CheckSubtitles()
-        {
-            var random = new Random();
-            Subtitles.Add(new FoundSubtitle
-            {
-                Checked = true,
-                SubtitleName = String.Format("subtitle {0}.srt", random.Next(0, 20))
-            });
-
-            Subtitles.Add(new FoundSubtitle
-            {
-                Checked = false,
-                SubtitleName = String.Format("subtitle {0}.srt", random.Next(0, 20))
-            });
-        }
-    }
-
-    public class FoundSubtitle
-    {
-        public string SubtitleName { get; set; }
-        public bool Checked { get; set; }
     }
 }
