+++ EasySubtitle.WPF/Windows/AdvancedSubtitlesWindow.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="EasySubtitle.WPF.AdvancedSubtitlesWindow"
+﻿<Window x:Class="EasySubtitle.WPF.Windows.AdvancedSubtitlesWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         Title="MainWindow" Height="500" Width="900">
