+++ EasySubtitle.WPF/Windows/ProgressDialogWindow.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="EasySubtitle.WPF.ProgressDialogWindow"
+﻿<Window x:Class="EasySubtitle.WPF.Windows.ProgressDialogWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         Title="ProgressDialogWindow" WindowStartupLocation="CenterOwner"
