+++ EasySubtitle.WPF/Windows/ProgressDialogWindow.xaml.cs
@@ -1,7 +1,7 @@
 ﻿using System.ComponentModel;
 using System.Windows;
 
-namespace EasySubtitle.WPF
+namespace EasySubtitle.WPF.Windows
 {
     /// <summary>
     /// Interaction logic for ProgressDialogWindow.xaml
